Bug fix: do not destroy user data on consent revoke

+++ app/controllers/sessions_controller.rb
@@ -6,8 +6,12 @@ class SessionsController < ApplicationController
       redirect_to '/consents/new'
     elsif !current_user.consents.first.username ||
           !current_user.consents.first.roles
-      current_user.destroy
-      redirect_to :logout
+      if current_user.character.nil?
+        current_user.destroy 
+        redirect_to :logout
+      else
+        redirect_to '/consents/new'
+      end
     elsif current_user.character.nil?
       current_user.build_character
       current_user.character.init(session[:user]['token'])
